+++ core/src/main/java/org/realityforge/arez/spy/ObservableChangedEvent.java
@@ -5,7 +5,6 @@ import java.util.Objects;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import org.realityforge.arez.Arez;
-import org.realityforge.arez.Observable;
 
 /**
  * Notification when Observable has changed.
@@ -16,18 +15,18 @@ public final class ObservableChangedEvent
   public static final String TYPE_NAME = EventUtil.getName( ObservableChangedEvent.class );
 
   @Nonnull
-  private final Observable<?> _observable;
+  private final ObservableInfo _observable;
   @Nullable
   private final Object _value;
 
-  public ObservableChangedEvent( @Nonnull final Observable<?> observable, @Nullable final Object value )
+  public ObservableChangedEvent( @Nonnull final ObservableInfo observable, @Nullable final Object value )
   {
     _observable = Objects.requireNonNull( observable );
     _value = value;
   }
 
   @Nonnull
-  public Observable<?> getObservable()
+  public ObservableInfo getObservable()
   {
     return _observable;
   }