+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### Unreleased
 
+##### Changed
+* 💥 **\[core\]** Introduce `ObservableInfo` interface that is used from within the spy subsystem. This change
+  effectively removes the `dispose()` method from the public interface of Observables accessed solely through
+  the spy subsystem.
+
 ### [v0.33](https://github.com/arez/arez/tree/v0.33) (2017-12-04)
 [Full Changelog](https://github.com/arez/arez/compare/v0.32...v0.33)